app_input: add save_with_name and record last-save / last-load readouts
Also bumps last_save_tick + RecentSaveRow.tick from u32 to u64 to match
sim.tick (plan typo — sim.tick has been u64 throughout).

+++ src/app.rs
@@ -325,7 +325,7 @@ pub(crate) struct AppState {
     /// Lives in AppState so the field persists across frames while open.
     pub(crate) dev_overlay_save_name: String,
     /// Tick number recorded by the most recent save this session.
-    pub(crate) last_save_tick: Option<u32>,
+    pub(crate) last_save_tick: Option<u64>,
     /// Wall-clock instant of the most recent save this session.
     pub(crate) last_save_instant: Option<std::time::Instant>,
     /// Path of the most recently loaded save (for "Reload last load").
